git-svn-id: http://opensvn.wiimm.de/wii/trunk/wiimms-iso-tools@2289 4ca24660-9720-11de-8e73-472ee046f03b

+++ setup/install.sh
@@ -30,7 +30,7 @@ echo "*** install binaries to $BIN_PATH"
 
 for f in $BIN_FILES
 do
-    install $INST_FLAGS bin/$f "$BIN_PATH/$f"
+    [[ -f bin/$f ]] && install $INST_FLAGS bin/$f "$BIN_PATH/$f"
 done
 
 for f in $WDF_LINKS
